requirements.txt
streamlit==0.88.0
Pillow==8.3.2
opencv-python-headless==4.9.0
numpy==1.21.2
matplotlib==3.5.1
tensorflow==2.8.0